add a web config creation script
